apps/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["AngeloCarnevale"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.0"
pyjwt = "^2.8.0"
django-cors-headers = "^4.2.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.0"}
ngrok = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
